refactor: 拆分并移动 WebSocket 管理器中的具体业务逻辑到新模块和新目录下
